Gravity working
Dirty gravity works, I'll need to do a proper delta time and currently
the interaction of matching of Sprite and APBody is more external.
However I now have the events and loop in place to focus on all the
collisions and movement. APBodyList is not complete however I'm ready
to push to github.