World constructor changed and body adding.
The world constructor now has pixelsToMetre as the first default because
I think this will be the one most likely to change instead of the fairly
common 9.8 m/s2 for gravity. Body adding is now done by creating a new
body object as a parameter that is then returned so it can be assigned
if needed and processed further all in one line.