Rotation works on APSprites
rotation now works correctly on APSprite objects, however I think there
should be an inverse to rotate the body if rotate is called DIRECTLY on
the APSprite. This may be added later but for now the rotating sprite
works correctly without needing matrix transforms. This will probably
change later but for now it functions.